Fixing vnet connections by adding constructor param for specifying domain
When using the constructor intended for use within a vnet by specifying numRegionServers, the load balancer was not using the fully qualified host name which caused name resolution errors in the client. Fixed by adding a new param to specify the vnet domain the cluster is on. Added a unit test to verify the fully qualified name is used, and manually verified on a cluster deployed within a vnet.