Fixed "Arithmetic operation resulted in an overflow." error
During normal operation I saw this error come up when running read
scans. The variable was an int and can overflow if incremented past max.
We start with a random value so it will be arbitrarily close to this max
at init time. The value is now a uint so that it will go to 0 and not
overflow when incremented past max.